10153 Move substantive reads from model package to transformation package in order to make it easier to exclude simple formats from scoverage
