+++ test/uk/gov/hmrc/tai/model/domain/CompanyCarBenefitSpec.scala
@@ -19,9 +19,9 @@ package uk.gov.hmrc.tai.model.domain
 import org.scalatestplus.play.PlaySpec
 import play.api.libs.json.{JsResultException, Json}
 import uk.gov.hmrc.domain.{Generator, Nino}
-import uk.gov.hmrc.tai.model.domain.benefits.CompanyCar.companyCarReadsFromHod
-import uk.gov.hmrc.tai.model.domain.benefits.CompanyCarBenefit.companyCarBenefitReadsFromHod
 import uk.gov.hmrc.tai.model.domain.benefits.{CompanyCar, CompanyCarBenefit}
+import uk.gov.hmrc.tai.transformation.CompanyCarBenefitTransformer.companyCarBenefitReadsFromHod
+import uk.gov.hmrc.tai.transformation.CompanyCarTransformer.companyCarReadsFromHod
 
 import java.time.LocalDate
 import scala.util.Random
